Scripts updated to run successfully in non-admin user accounts

+++ scripts/99.optimize-second-users-ui-only.ps1
@@ -4,15 +4,9 @@
 # ===========
 # Runs only the script that optimize user UI for second users
 
-Import-Module -DisableNameChecking $PSScriptRoot\..\lib\common-lib.psm1
+Import-Module -DisableNameChecking $PSScriptRoot\..\lib\common-lib.psm1 -Force
 
-# Preparation
-#============================================================================
-
-Check-AdminRights
-
-# Real work
-#=============================================================================
+$username = Get-LoggedUsername
 
 Print-Message-With-Banner("Starting User UI Cleanup")
 
@@ -20,16 +14,18 @@ Print-Message-With-Banner("Starting User UI Cleanup")
 &($PSScriptRoot+"\6.1.cleanup-taskbar.ps1")
 &($PSScriptRoot+"\7.unbloat-start-menu.ps1")
 
-if (Is-BuiltInAdminLoggedInUser) {
-    Write-Debug "Built-in admin account in session - User demotion skipped"
-    Return
-} else {
+if (Is-UserAdministrator $(Get-LoggedUsername)) {
+    if (Is-BuiltInAdminLoggedInUser) {
+        Write-Debug "Built-in admin account in session - User demotion skipped"
+        Return
+    } else {
 
-    $choice = [Microsoft.VisualBasic.Interaction]::MsgBox('Remove current user from Administrators group?', 'YesNo,SystemModal,Question', 'Remove user from admin group check')
+        $choice = [Microsoft.VisualBasic.Interaction]::MsgBox('Remove current user from Administrators group?', 'YesNo,SystemModal,Question', 'Remove user from admin group check')
 
-    switch  ($choice) {
-    'Yes' {
-        Remove-CurrentUserAdminGroup
-	    }
+        switch  ($choice) {
+        'Yes' {
+            Remove-CurrentUserAdminGroup
+	        }
+        }
     }
 }
